requirements.txt
# Core pipeline
numpy>=1.24.0
pdfminer.six
pdf2image
Pillow>=10.0.0
pytesseract>=0.3.10

# Data validation
pydantic>=2.0.0

# LLM and HTTP
requests
python-dotenv
loguru

# Database
sqlalchemy>=2.0.0

# CLI
typer

# API (optional, for Docker deployment)
fastapi==0.109.0
uvicorn[standard]>=0.31.1
python-multipart>=0.0.9

# Datasets (Hugging Face)
datasets>=2.16.0
huggingface_hub>=0.20.0

# Assistant service (MCP)
mcp @ git+https://github.com/modelcontextprotocol/python-sdk.git
